AI Summarizer Working

+++ backend/src/main/java/com/cognitum/backend/properties/JwtProperties.java
@@ -13,6 +13,6 @@ public class JwtProperties {
 
     private String secretKey;
     private Integer jwtExpiration;
-    private Integer refreshTokenExpiration;
+    private Long refreshTokenExpiration;
 
 }
